DMS-Changes
New Attributes:
- cis_suche: Defines if file is searchable via CIS
- schlagworte: Semikolon-Separated Text for CIS-Search
- kategorie_oe_kurzbz: Organisational Unit defines rights for viewing, uploading and changing documents in DMS
- berechtigung_kurzbz: Controls visibility for categories in DMS

Access to categories can now be restricted with oe_kurzbz and rights. If an Organisational Unit is added to a category, only users with the according rights can access, upload and change documents in this category.

If a right is saved to a category it is only accessible for users with this right. Use the right to restrict access to sensible DMS-Categories (e.g. FAS-Documents)

!! Downloading of documents via dms.php is still controlled only with the group assigned to the category !!

New right "basis/dmsAdmin" needed for accessing DMS-Administration (Editing categories)