Added Reihungstest-Gebiete for person
Now the areas for the Reihungstest are put together depending on the studies the examinee is interested in and if the examinee is a Quereinsteiger or not.

Therefore all prestudents of the examinee are detected that are set for the actual and future studiensemester. 

The areas are chosen because of the Studienplan. If not available, they are chosen because of the Studiengang.