Separated Bachelor and Masters on testtool
Now if the logged-in prestudents study is a Bachelor-study, all other Bachelor studies he might have applied for, are displayed. Also the testing areas (gebiete) are shown only for that Bachelor studies.

If the logged-in prestudents study is a Master (or something else than a Bachelor), only that one particular Master study is displayed on the starting page.  The testing areas are shown only for that one particular Master study.