Fixed to get question in language of the actual logged-in Master study
When logging in with a Master study, the system was retrieving the
studyplan/study of the highest priored applied Master study. This
behaviour should only occure for Bachelor studies and was leading to
incorrect language setting.

Now if user logs in with a Master study, the actual studyplan/study and
the correct language setting are retrieved.