Adapted Controller to UPDATE also CHANGED HOURS
Up to now changes were identified only by comparing betrag with
vertrag_betrag.
Problem is: when a new Lehrauftrag betrag is set with 0 or NULL and only
hours are set, a later change of hours cannot be identified (as the betrag
and vertrag_betrag do not change). This is fixed now.