Fixed & Secured: Approving LA not causing permission error anymore
Before the organisational unit of the lv was sent via AJAX to the server
and used for checking the permissions on that organisational unit.
This was insecure (post data manipulation) and organisational units were
wrong (prepared string instead of oe_kurzbz).

This is fixed now by serverside check of organisational unit.
(based on given vertrag_id)