Added sending Sanchomail to lector asking to recommend Anrechnung
Mails are send to lectors (prio for LV-Leitung, if not present to all
lectors of LV) Anyway, as many there recommendation may be requested
for many lvs at the same time, the receiver array will be unique to ensure
sending only once to one and the same lector. e.g. if lector is lv-leitung
for lv x and lector for lv y, will get only one mail.

Signed-off-by: cris-technikum <hainberg@technikum-wien.at>